Add files via upload
Original FreeCAD of liner and case (PDM_Stator_Pipe1).  The origin of the housing looks to be in the middle of the nipple, and the liner code grok3 gave is offset by half the housing.  The second attempt of the code (PDM_Stator_Pipe2) aligned the liner with the housing but the internal dimensions of the liner are not concentric with the housing at the ends of the housing.  The non-concentric ends may cause problems in the future with bearings.  The OBJ files are included for the housing and lining.